refactor: create rows in blueprint middleware
